pyproject.toml
[project]
name = "programgarden"
version = "0.2.0"
description = "LS증권 API을 메인으로 제작된 노코딩 시스템 트레이딩 DSL 오픈소스"
authors = [
    {name = "장용준",email = "coding@programgarden.com"}
]
homepage = "https://programgarden.com"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["LS증권", "finance", "trading"]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
requests = ">=2.32.3,<3.0.0"
pydantic = ">=2.11.4,<3.0.0"
python-dotenv = ">=1.1.0,<2.0.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"
pep8-naming = "^0.15.0"
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
requests-mock = "^1.12.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
